SpreadsheetHttpServerHttpHandler: SpreadsheetServerContext dependency removed
